% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLD.R
\name{generateLD}
\alias{generateLD}
\title{Calculate LD files}
\usage{
generateLD(assoc, path, bfile_path, cores = 10, ethnic = "eur",
  plink_path = NULL, python_path = NULL)
}
\arguments{
\item{assoc}{GWAS summary statistics, need to include snp, chr, a0, a1, z (header is necessary)}

\item{path}{The output path}

\item{bfile_path}{The path to the plink bfiles for LD estimation}

\item{cores}{The number of cores for computation in parallel}

\item{ethnic}{Ethnic of the GWAS cohort; 'eur' for European ancestry}

\item{plink_path}{The path to the plink software. If not specified, the function will use the default path (which plink)}

\item{python_path}{The path to the python software. If not specified, the function will use the default path (which python)}
}
\description{
Run LDER
}
