% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLDER.R
\name{runLDER}
\alias{runLDER}
\title{Main function}
\usage{
runLDER(assoc, n.gwas, path, LD.insample = F, n.ld, ethnic = "eur",
  method = "lder", cores = 10, a = NULL, rough = T, twostage = T,
  type = "jack", n.bs = 1000)
}
\arguments{
\item{assoc}{GWAS summary statistics, need to include snp, chr, a0, a1, z (header is necessary)}

\item{n.gwas}{The sample size of the GWAS summary statistics}

\item{path}{The output path}

\item{LD.insample}{T/F, whether the LD reference is estimated with the GWAS cohort (T) or external reference panel (e.g. 1000 Genome Project) (F)}

\item{n.ld}{The sample size of the LD reference}

\item{ethnic}{Ethnic of the GWAS cohort; 'eur' for European ancestry}

\item{method}{'lder', 'ldsc', or 'both'}

\item{cores}{The number of cores for computation in parallel}

\item{a}{Pre-specified inflation factor, default=NULL}
}
\description{
Run LDER
}
